"password update"

+++ graphql-bootcamp/graphql-prisma/src/resolvers/Mutation.js
@@ -1,14 +1,11 @@
 import bcrypt from 'bcryptjs'
-import jwt from 'jsonwebtoken'
 import getUserId from '../utils/getUserId'
+import generateToken from '../utils/generateToken'
+import hashPassword from '../utils/hashPassword'
 
 const Mutation = {
     async createUser(parent, args, { prisma }, info) {
-        if (args.data.password.length < 8) {
-            throw new Error('Password must be 8 characters or longer.')
-        }
-
-        const password = await bcrypt.hash(args.data.password, 10)
+        const password = await hashPassword(args.data.password)
         const user = await prisma.mutation.createUser({
             data: {
                 ...args.data,
@@ -18,7 +15,7 @@ const Mutation = {
 
         return {
             user,
-            token: jwt.sign({ userId: user.id }, 'thisisasecret')
+            token: generateToken(user.id)
         }
     },
     async login(parent, args, { prisma }, info) {
@@ -40,7 +37,7 @@ const Mutation = {
 
         return {
             user,
-            token: jwt.sign({ userId: user.id }, 'thisisasecret')
+            token: generateToken(user.id)
         }
     },
     async deleteUser(parent, args, { prisma, request }, info) {
@@ -54,6 +51,10 @@ const Mutation = {
     },
     async updateUser(parent, args, { prisma, request }, info) {
         const userId = getUserId(request)
+
+        if (typeof args.data.password === 'string') {
+            args.data.password = await hashPassword(args.data.password)
+        }
 
         return prisma.mutation.updateUser({
             where: {
@@ -97,7 +98,24 @@ const Mutation = {
             }
         }, info)
     },
-    updatePost(parent, args, { prisma }, info) {
+    async updatePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+        const isPublished = await prisma.exists.Post({ id: args.id, published: true })
+
+        if (!postExists) {
+            throw new Error('Unable to update post')
+        }
+
+        if (isPublished && args.data.published === false) {
+            await prisma.mutation.deleteManyComments({ where: { post: { id: args.id } } })
+        }
+
         return prisma.mutation.updatePost({
             where: {
                 id: args.id
@@ -105,13 +123,23 @@ const Mutation = {
             data: args.data
         }, info)
     },
-    createComment(parent, args, { prisma }, info) {
+    async createComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists = await prisma.exists.Post({
+            id: args.data.post,
+            published: true
+        })
+
+        if (!postExists) {
+            throw new Error('Unable to find post')
+        }
+
         return prisma.mutation.createComment({
             data: {
                 text: args.data.text,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -122,14 +150,38 @@ const Mutation = {
             }
         }, info)
     },
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to delete comment')
+        }
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, args, { prisma }, info) {
+    async updateComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if (!commentExists) {
+            throw new Error('Unable to update comment')
+        }
+
         return prisma.mutation.updateComment({
             where: {
                 id: args.id
